make sure $inc lib is prefixed with -I (RT#73659)
For some reason using the inc Module::Install command to specify the -I
options to the compiler was failing, the $(INC) was being passed to the
compiler with the -I stripped in the Makefile.

Change this to makemaker_args CCFLAGS => $inc and now we get the the
proper -I options; unfortunately the bare INC value is still present,
but does not seem to be doing any harm yet. I will try to think of a way
to get rid of it. This seems to be some sort of EUMM bug.