Fix auth_keyboard file descriptor leaking
I have fully rewritting the code for this function. IMO, now it is much
easier to understand.

Instead of the sv_tmp field on the session wrapping structure now it
uses a localized global. That allows us to handle the scope of this data
in a sanner and safer way. For instance, dieing from the callback is
now safe.

The down side of this is that it is still hairy code (all this cb stuff
always is) and I have only being able to tests some of the possible cases.
Specifically, I don't have access to any server implementing a pasword
expiration policy.