Generate constant groups automatically
Until now, constants had to be maintained on two places, one the list
in Makefile.PL given to ExtUtils::Constant and the other inside
lib/Net/SSH2.pm, on the export tags definitions. That was quite
annoying and error prone.

In this patch, all the constant handling is moved to a new script at
util/gen_constants.pl which generates const-xs.inc, const-c.inc and
lib/Net/SSH2/Constants.pm.

The script is expected to be called by module authors only and is not
included on the distribution.

Also, the old fallback directory containing pregenerated versions of
const-*.inc, just in case the user didn't had ExtUtils::Constant
installed, has also been removed.