READLINE: remove hard-coded timeout
READLINE had a hard-coded timeout of 250ms which made it unreliable as
data can be delayed far longer by the network layer, or the remote
program could generate data slower than that.

This change is backward incompatible and will cause scripts depending
on that timeout to hang. But those scripts were broken anyway.

This change adds also support for the new "readline" and "getc"
methods that are just extended versions of the old "READLINE" and
"GETC" allowing to choose the reading string (ext) and the end of line
marker.