readline and getc: try to detect EOF better.
Due to the SSH protocol sharing the EOF between the two incoming
streams, it is not possible to detect if all data from any of them has
been completely read when data remains still queued in the other one.

That makes the "readline" method still unreliable.

Anyway, this patch can detect the EOF condition in more cases.

Also, prepare for release 0.59_13.