Handle stream_id arguments using a typemap
The wrapping code was using a macro for converting stream_id arguments
that would only produce 0 or 1.

Even if in practice that was enough, as no common implementation of
SSH protocol uses anything besides those ids, libssh2 support any id,
so the wrapper should behave in the same way.

Also, in flush some magic values have special meanings and those were
being handled as 1.

With this patch, stream_id arguments are handled as constants and
special values are converted to integers properly.