Warn about readline being called in non-blocking mode
Calling Net::SSH2::Channel::readline in non-blocking mode is a common
error in user code. Present in several code snippets and samples
available from the Internet in sites like Perlmonks if has become a
cargo-cult issue.

People also uses it when talking to a remote shell in order to read
the output from some command before sending a new one. A method that
works most of the time but that we should disregard as it is completely
unreliable.

This patch just makes the module emit a warning when such situation is
detected.