remove extra test for deadline
This line follows the same test four lines above.

At best, this line does nothing, since it is only tested if the previous
test failed, so this test will almost always fail too.

At worst, the timing will be exactly right for the first test to fail
and the second test to pass, which would result in a return without
setting the "Time out" error.